requirements.txt
langchain
openai
wikipedia
chromadb
tiktoken
python-dotenv